Fixes for nginx 1.20.2+ (#4)
* Add 1.21.5,1.22.0 and 1.23.0 to the CI
* Fixed regex parsing on NGINX 1.21.5 or newer
* Fixed build on 1.23.0
* Check that debug logs are disabled.
* Fixed tests due newer NGINX parser

+++ .scripts/ci-build.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 # SPDX-FileCopyrightText: 2022 wargio <deroad@libero.it>
 # SPDX-License-Identifier: LGPL-3.0-only
+set -e
 
 NGINX_VERSION="$1"
 N_CPUS=$(nproc)
@@ -15,38 +16,48 @@ echo "############################"
 echo "   NGINX VERSION: $NGINX_VERSION"
 echo "############################"
 
-rm -rf nginx-source nginx-tmp nginx.tar.gz 2>&1 > /dev/null
+NEW_BUILD=true
+if [ -d "nginx-tmp" ] && [ "$NGINX_VERSION" == $(cat nginx-tmp/nginx.version) ]; then
+    NEW_BUILD=false
+fi
 
-set -e
+if $NEW_BUILD ; then
+    rm -rf nginx-source nginx-tmp nginx.tar.gz 2>&1 > /dev/null
+    wget --no-clobber -O nginx.tar.gz "https://nginx.org/download/nginx-$NGINX_VERSION.tar.gz"
+    mkdir -p nginx-source nginx-tmp/naxsi_ut/root
+    echo "$NGINX_VERSION" > nginx-tmp/nginx.version
+    tar -C nginx-source -xzf nginx.tar.gz --strip-components=1
+    rm nginx.tar.gz
+fi
 
-wget --no-clobber -O nginx.tar.gz "https://nginx.org/download/nginx-$NGINX_VERSION.tar.gz"
-mkdir -p nginx-source nginx-tmp/naxsi_ut/root
-tar -C nginx-source -xzf nginx.tar.gz --strip-components=1
-rm nginx.tar.gz
 export NAXSI_CFG_PATH=$(realpath naxsi_config/)
 export NAXSI_SRC_PATH=$(realpath naxsi_src/)
 export NAXSI_TMP_PATH=$(realpath nginx-tmp/)
 export NAXSI_TST_PATH=$(realpath unit-tests/)
 export NGINX_TMP_PATH=$(realpath nginx-source/)
 export CFLAGS="-Wextra -Wall" # -Werror"
-cd nginx-source
-./configure --with-select_module \
-            --conf-path="$NAXSI_TMP_PATH/naxsi_ut/nginx.conf" \
-            --http-client-body-temp-path="$NAXSI_TMP_PATH/naxsi_ut/body/" \
-            --http-fastcgi-temp-path="$NAXSI_TMP_PATH/naxsi_ut/fastcgi/" \
-            --http-proxy-temp-path="$NAXSI_TMP_PATH/naxsi_ut/proxy/" \
-            --lock-path="$NAXSI_TMP_PATH/nginx.lock" \
-            --pid-path="$NAXSI_TMP_PATH/naxsi_ut/nginx.pid" \
-            --modules-path="$NAXSI_TMP_PATH/naxsi_ut/modules/" \
-            --without-mail_pop3_module \
-            --without-mail_smtp_module \
-            --without-mail_imap_module \
-            --with-http_v2_module \
-            --without-http_uwsgi_module \
-            --without-http_scgi_module \
-            --prefix="$NAXSI_TMP_PATH/" \
-            --add-dynamic-module="$NAXSI_SRC_PATH" \
-            --error-log-path="$NAXSI_TMP_PATH/naxsi_ut/error.log" \
-            --conf-path="$NAXSI_TMP_PATH/naxsi_ut/nginx.conf"
-cd ..
+
+if $NEW_BUILD ; then
+    cd "$NGINX_TMP_PATH"
+    ./configure --with-select_module \
+                --conf-path="$NAXSI_TMP_PATH/naxsi_ut/nginx.conf" \
+                --http-client-body-temp-path="$NAXSI_TMP_PATH/naxsi_ut/body/" \
+                --http-fastcgi-temp-path="$NAXSI_TMP_PATH/naxsi_ut/fastcgi/" \
+                --http-proxy-temp-path="$NAXSI_TMP_PATH/naxsi_ut/proxy/" \
+                --lock-path="$NAXSI_TMP_PATH/nginx.lock" \
+                --pid-path="$NAXSI_TMP_PATH/naxsi_ut/nginx.pid" \
+                --modules-path="$NAXSI_TMP_PATH/naxsi_ut/modules/" \
+                --without-mail_pop3_module \
+                --without-mail_smtp_module \
+                --without-mail_imap_module \
+                --with-http_v2_module \
+                --without-http_uwsgi_module \
+                --without-http_scgi_module \
+                --prefix="$NAXSI_TMP_PATH/" \
+                --add-dynamic-module="$NAXSI_SRC_PATH" \
+                --error-log-path="$NAXSI_TMP_PATH/naxsi_ut/error.log" \
+                --conf-path="$NAXSI_TMP_PATH/naxsi_ut/nginx.conf"
+    cd ..
+fi
+
 make -C "$NGINX_TMP_PATH" -j$N_CPUS install
